Refactored SodaFolderRunner and SodaFileRunner to conform to Command-Query separation.
Practice what you preach!
Rather than executing a test, writing a file and returning the result, SodaFileRunner now just executes and returns the result.
File writing is done by the SodaFolderRunner and the main() method of this no longer executes and returns, but instead returns Unit.
The exit() action that is only to be performed when main() is called from the command-line is now passed into the other main() as a function callback.